third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-allow-same-suborigin-access.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Allow suborigin in HTTP header</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
var tests;
function next_test() {
  var test = tests.shift();
  if (test)
    test();
}

var tests = [
  function() {
    async_test(function(t) {
        window.onmessage = t.step_func(function(event) {
            assert_equals(event.data, 'I am a secret');
            t.done();
            next_test();
          });
        var iframe = document.createElement('iframe');
        iframe.src = 'resources/iframe-reaches-into-parent.php?' +
                     'suborigin=foobar&childsuborigin=foobar';
        document.body.appendChild(iframe);
      }, 'iframe reaches into parent')
  },
  function() {
    async_test(function(t) {
        window.onmessage = t.step_func(function(event) {
            assert_equals(event.data, 'I am a secret');
            t.done();
            next_test();
          });
        var iframe = document.createElement('iframe');
        iframe.src = 'resources/reach-into-iframe.php?' +
                     'suborigin=foobar&childsuborigin=foobar';
        document.body.appendChild(iframe);
      }, 'Parent reaches into iframe')
  }
];

next_test();
</script>
</body>
</html>
